refactor: move tenant schema to the tenant plugin
