+++ wazo_auth/plugins/http/tenants/tests/test_schemas.py
